Marks two methods as obsolete, and implements a compatibility layer for them. Better precompiled macro handling Moves IsCurrentPlatformSupported to an static member of the class to instantiate. Removes assembly to reduce plugin complexity
