+++ AppleAuth/AppleAuthManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3577fa6358a844916b45d108a7e7e26a
+guid: f13d47bb9704943a89585d5880141845
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
