+++ AppleAuth/IAppleAuthManager.cs
@@ -6,10 +6,14 @@ namespace AppleAuth
 {
     public interface IAppleAuthManager
     {
-        bool IsCurrentPlatformSupported { get; }
-
+        [Obsolete("This method is deprecated and will be removed soon. Please provide an empty instance of AppleAuthQuickLoginArgs to QuickLogin.")]
+        void QuickLogin(Action<ICredential> successCallback, Action<IAppleError> errorCallback);
+        
         void QuickLogin(AppleAuthQuickLoginArgs quickLoginArgs, Action<ICredential> successCallback, Action<IAppleError> errorCallback);
 
+        [Obsolete("This method is deprecated and will be removed soon. Please provide an instance of AppleAuthLoginArgs to LoginWithAppleId with the LoginOptions instead.")]
+        void LoginWithAppleId(LoginOptions options, Action<ICredential> successCallback, Action<IAppleError> errorCallback);
+        
         void LoginWithAppleId(AppleAuthLoginArgs loginArgs, Action<ICredential> successCallback, Action<IAppleError> errorCallback);
 
         void GetCredentialState(string userId, Action<CredentialState> successCallback, Action<IAppleError> errorCallback);
