--- AppleAuth/IOS/AppleAuth.IOS.asmdef
@@ -1,17 +1,0 @@
-{
-    "name": "AppleAuth.IOS",
-    "references": [
-        "AppleAuth"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [
-        "Editor",
-        "iOS"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": []
-}
