+++ AppleAuthSampleProject/Assets/Plugins/Editor.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 63285876e96e74d4a88b51cb8146f562
-AssemblyDefinitionImporter:
+guid: a1ec445199fbf4acbaaf4b8019e23053
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
